训练粗分类器

+++ utils/make_train_classes.py
@@ -21,7 +21,7 @@ def fit_one_epoch_classes(model, criterion, optimizer, item, max_epoch,
         # 训练
         for iteration, (image_tensor, target_t) in enumerate(train_loader):
             image_tensor = image_tensor.type(torch.FloatTensor).to(device)
-            target_t = target_t.to(device)
+            target_t = target_t.long().to(device)
             feature = model(image_tensor).to(device)
             loss = criterion(feature, target_t).to(device).mean()
             Loss += loss
